requirements.txt
fastapi
gunicorn
uvicorn
jinja2
pytest
requests
pre-commit
python-dotenv
aiofiles
python-multipart
pillow
pytesseract
PyMuPDF